test-requirements.txt
pytest==7.3.1
black==23.3.0
pylint==2.17.4
isort==5.12.0
mypy==1.2.0
sqlfluff==2.1.0
pandas-stubs==2.0.1.230501